pyproject.toml
[project]
name = "mlflow-or-crash-course"
version = "0.1.0"
description = "Crash course for benchmarking and deploying OR solvers with MLflow on Databricks"
readme = "README.md"
authors = [
    { name = "josh-melton-db", email = "josh-melton-db@users.noreply.github.com" }
]
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk>=0.103.0",
    "mlflow>=3.11.1",
    "numpy>=2.2.6",
    "ortools>=9.15.6755",
    "pandas>=2.3.3",
    "scipy>=1.15.3",
]

[project.scripts]
mlflow-or-crash-course = "mlflow_or_crash_course.cli:main"

[build-system]
requires = ["uv_build>=0.9.27,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
